requirements.txt
Django>=5.0,<6.0
djangorestframework
drf-yasg
djangorestframework-simplejwt